requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
exceptiongroup==1.3.0
frozenlist==1.7.0
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
multidict==6.6.4
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyTelegramBotAPI==4.28.0
pytest==8.4.1
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
schedule==1.2.2
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.20.1
pydub==0.25.1
openai==1.3.3